Working towards compiling with rust alpha

+++ src/db/lru_set.rs
@@ -4,7 +4,7 @@ use std::hash;
 
 use std::collections::HashSet;
 use std::collections::LruCache;
-use std::collections::hash_map::{Occupied, Vacant};
+// use std::collections::hash_map::{Occupied, Vacant};
 
 use db::store::Store;
 
@@ -13,7 +13,7 @@ pub struct LruSet<K, V> {
 }
 
 impl<K: hash::Hash + cmp::Eq + clone::Clone, V: hash::Hash + cmp::Eq + clone::Clone> LruSet<K, V> {
-    pub fn with_capacity(i: uint) -> LruSet<K, V> {
+    pub fn with_capacity(i: u64) -> LruSet<K, V> {
         let data: LruCache<K, HashSet<V>> = LruCache::new(i);
         return LruSet {data: data};
     }
